Change list separators
This comment makes semicolon mandatory. This makes the BNF easier to
read and it's closer to the standard. Optional semicolon rules are
expected to be implemented in the parser/lexer directly.

This comment also introduces trailing commas, with similar rules as the
semicolon. For example:

	function foo(
		int p1,
		int p2,  // trailing comma
	) return Point
	{
		return {
			x := p1,
			y := p2, // trailing comma
		}
	}

This trailing comma is not standardized, yet. But it probably will be
and it's convenient.

Also, some minor changes have been introduced:
* add missing Block-reference in else-ifs
* replace ValueConstraints with "(" Exprs ")"